Fix review detail INT4 decode and enrich command ACL plus UI.
Align findings line/column types with Postgres INTEGER so GET /reviews/:id stops 500ing, widen who can run bot commands with a denial reply, and expand the review detail page with KPIs, filters, and richer finding cards.

+++ src/bot/review/persist.rs
@@ -65,7 +65,7 @@ pub(crate) async fn save_review_to_db(
             fingerprint: Some(format!("{}:{}", review.id, f.fingerprint())),
             file_path: f.file.clone(),
             line_start: if f.line > 0 {
-                Some(f.line as i64)
+                Some(f.line as i32)
             } else {
                 None
             },
